# piletimyyk-api/docs/swagger.yaml
openapi: 3.0.2
info:
    title: Piletimyyk-API
    version: 1.0.0
    description: API for the sale and management of event tickets
paths:
...
    /users/login:
        post:
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/user'
                        examples:
                            User Credentials:
                                value:
                                    -
                                        email: user@example.com
                                        password: password12345
                required: true
            tags:
                - users
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/message'
                            examples:
                                Login Success Response:
                                    value:
                                        -
                                            message: Login successful.
                    description: Confirms that user has successfully logged in.
                '401':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/message'
                            examples:
                                Invalid Credentials Response:
                                    value:
                                        -
                                            message: ' Invalid email or password.'
                    description: >-
                        User should check their credentials and try again, because the provided email or
                        password is incorrect.
                '500':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/message'
                            examples:
                                Internal Server Error Response:
                                    value:
                                        message: Error logging in.
                    description: >-
                        User should try again later or contact support if the issue of an error on the server
                        while processing the request persists.
            deprecated: false
            operationId: loginUser
            summary: User login
            description: User to be authenticated.
components:
    schemas:
...
        user:
            title: Root Type for user
            description: ''
            type: object
            properties:
                user_id:
                    format: int32
                    type: integer
                email:
                    type: string
                first_name:
                    type: string
                last_name:
                    description: ''
                    type: string
            example:
                user_id: 42
                first_name: Krista
                last_name: Tamm
                email: kristat@example.com
...
        message:
            title: Root Type for message
            description: ''
            type: object
            properties:
                message:
                    type: string
            example:
                message: Login successful